Fixes per cppcheck; made translation_record have no vtable (saves space)
Mostly initializing class members and passing by reference.

+++ Felix/AdvancedMemMgrDlg.h
@@ -47,7 +47,7 @@ public:
 	{}
 
 	int get_validation() const;
-	int get_reliability() const;
+	size_t get_reliability() const;
 	int get_lock() const;
 
 	void set_validation(int setting);
